Chop everything up into little modules; fix copy-paste bug in account creation
