+++ app/create-account.html
@@ -19,6 +19,6 @@
         
         <script src="./src/third/pdf.js.build/pdf.js"></script>
         <script src="https://cdn.jsdelivr.net/npm/appwrite@6.0.1"></script>
-        <script type="module" src="./src/create-account.js"></script>
+        <script type="module" src="./src/screens/create-account/ui/create-account.js"></script>
     </body>
 </html>
