+++ app/dashboard.html
@@ -30,6 +30,6 @@
 
         <script src="./src/third/pdf.js.build/pdf.js"></script>
         <script src="https://cdn.jsdelivr.net/npm/appwrite@6.0.1"></script>
-        <script type="module" src="./src/dashboard.js"></script>
+        <script type="module" src="./src/screens/dashboard/ui/dashboard.js"></script>
     </body>
 </html>
