+++ app/index.html
@@ -213,6 +213,6 @@
         
         <script src="./src/third/pdf.js.build/pdf.js"></script>
         <script src="https://cdn.jsdelivr.net/npm/appwrite@6.0.1"></script>
-        <script type="module" src="./src/index.js"></script>
+        <script type="module" src="./src/screens/editor/ui/editor.js"></script>
     </body>
 </html>
