+++ app/login.html
@@ -19,6 +19,6 @@
         
         <script src="./src/third/pdf.js.build/pdf.js"></script>
         <script src="https://cdn.jsdelivr.net/npm/appwrite@6.0.1"></script>
-        <script type="module" src="./src/login.js"></script>
+        <script type="module" src="./src/screens/login/ui/login.js"></script>
     </body>
 </html>
