+++ app/src/common/model/appwrite.js
@@ -1,0 +1,6 @@
+var appwrite = new Appwrite();
+appwrite
+    .setEndpoint("https://appwrite.software-engineering.education/v1")
+    .setProject(/* crity */ "6206644928ab8835c77f");
+
+export {appwrite};
