+++ app/src/screens/dashboard/model/data.js
@@ -1,0 +1,13 @@
+import { AccountSession } from "../../../common/model/AccountSession.js";
+import { PresentationList } from "../../../common/model/PresentationList.js";
+
+class DashboardData {
+    constructor() {
+        this.accountSession = new AccountSession();
+        this.presentationList = new PresentationList(this.accountSession);
+    }
+}
+
+var data = new DashboardData();
+
+export { data, DashboardData };
