+++ app/src/screens/editor/model/Version.js
@@ -1,0 +1,9 @@
+class Version {
+    constructor(label, pdfUrl, appwriteId) {
+        this.label = label;
+        this.pdfUrl = pdfUrl;
+        this.appwriteId = appwriteId;
+    }
+}
+
+export { Version };
